fix up incorrect priv key storage
