+++ library/src/androidTest/java/org/xmtp/android/library/InstrumentedTest.kt
@@ -215,95 +215,4 @@ class InstrumentedTest {
 //            XCTFail("no messages")
         }
     }
-
-    @Test
-    fun testCanReadGzipCompressedMessages() {
-        val ints = arrayOf(
-            225, 2, 36, 98, 37, 243, 68, 234,
-            42, 126, 248, 246, 126, 83, 186, 197,
-            204, 186, 19, 173, 51, 0, 64, 0,
-            155, 8, 249, 247, 163, 185, 124, 159,
-        )
-        val keyBytes =
-            ints.foldIndexed(ByteArray(ints.size)) { i, a, v -> a.apply { set(i, v.toByte()) } }
-
-        val key = PrivateKeyOuterClass.PrivateKey.newBuilder().also {
-            it.secp256K1Builder.bytes = keyBytes.toByteString()
-            it.publicKeyBuilder.secp256K1UncompressedBuilder.bytes =
-                KeyUtil.addUncompressedByte(KeyUtil.getPublicKey(keyBytes)).toByteString()
-        }.build()
-
-
-        val client = Client().create(account = PrivateKeyBuilder(key))
-        assertEquals(client.apiClient.environment, XMTPEnvironment.DEV)
-        val convo = client.conversations.list()[0]
-        val message = convo.messages()[0]
-        assertEquals("hello gzip", message.content())
-    }
-
-    @Test
-    fun testCanReadZipCompressedMessages() {
-        val ints = arrayOf(
-            60, 45, 240, 192, 223, 2, 14, 166,
-            122, 65, 231, 31, 122, 178, 158, 137,
-            192, 97, 139, 83, 133, 245, 149, 250,
-            25, 125, 25, 11, 203, 97, 12, 200,
-        )
-        val keyBytes =
-            ints.foldIndexed(ByteArray(ints.size)) { i, a, v -> a.apply { set(i, v.toByte()) } }
-
-        val key = PrivateKeyOuterClass.PrivateKey.newBuilder().also {
-            it.secp256K1Builder.bytes = keyBytes.toByteString()
-            it.publicKeyBuilder.secp256K1UncompressedBuilder.bytes =
-                KeyUtil.addUncompressedByte(KeyUtil.getPublicKey(keyBytes)).toByteString()
-        }.build()
-
-        val client = Client().create(account = PrivateKeyBuilder(key))
-        assertEquals(client.apiClient.environment, XMTPEnvironment.DEV)
-        val convo = client.conversations.list()[0]
-        val message = convo.messages().lastOrNull()!!
-        val kotlindata = "hello deflate"
-        print("kotlin version: ${kotlindata.zlibCompress()}")
-        assertEquals("hello deflate", message.content())
-        convo.send(
-            text = "hello deflate from kotlin again",
-            SendOptions(compression = EncodedContentCompression.DEFLATE))
-    }
-
-    fun String.zlibCompress(): ByteArray {
-        val input = this.toByteArray(charset("UTF-8"))
-
-        // Compress the bytes
-        // 1 to 4 bytes/char for UTF-8
-        val output = ByteArray(input.size * 4)
-        val compressor = Deflater().apply {
-            setInput(input)
-            finish()
-        }
-        val compressedDataLength: Int = compressor.deflate(output)
-        return output.copyOfRange(0, compressedDataLength)
-    }
-
-    @Test
-    fun testCanLoadAllConversations() {
-        val ints = arrayOf(
-            105, 207, 193, 11, 240, 115, 115, 204,
-            117, 134, 201, 10, 56, 59, 52, 90,
-            229, 103, 15, 66, 20, 113, 118, 137,
-            44, 62, 130, 90, 30, 158, 182, 178,
-        )
-        val keyBytes =
-            ints.foldIndexed(ByteArray(ints.size)) { i, a, v -> a.apply { set(i, v.toByte()) } }
-
-        val key = PrivateKeyOuterClass.PrivateKey.newBuilder().also {
-            it.secp256K1Builder.bytes = keyBytes.toByteString()
-            it.publicKeyBuilder.secp256K1UncompressedBuilder.bytes =
-                KeyUtil.addUncompressedByte(KeyUtil.getPublicKey(keyBytes)).toByteString()
-        }.build()
-        val client = Client().create(account = PrivateKeyBuilder(key))
-        val conversations = client.conversations.list()
-        assertEquals(200, conversations.size)
-    }
-
-
 }
